EEVEE: Film: Add ViewData that have panoramic view direction infomation for panoramic projection ------------
This commit has been modified to project the 6 cubes view onto the Film with Equirectangular.
In order to project 6 cubes, we need to send the Direction information of the View to the shader.
In this commit, we define a new struct called ViewData and send it to the shader.
An alternative change would be to add the direction information to CameraData and FilmData, which are already sent to the shader. However, since it is semantically associated with the ShadingView, we have added it as a new member variable to the ShadingView.

Differential Revision: https://developer.blender.org/D13536